REFACTORING: Moved node loading to separate file
