+++ cms_named_menus/admin.py
@@ -2,11 +2,9 @@ import json
 
 from django.conf import settings
 from django.contrib import admin
-from django.contrib.auth.models import AnonymousUser
-from menus.menu_pool import menu_pool
 
-from .models import CMSNamedMenu
-
+from cms_named_menus.models import CMSNamedMenu
+from cms_named_menus.nodes import get_nodes
 
 class LazyEncoder(json.JSONEncoder):
     """Encodes django's lazy i18n strings.
@@ -18,52 +16,29 @@ class LazyEncoder(json.JSONEncoder):
             return force_unicode(obj)
         return obj
 
-def anonymous_request(f):
-    def decorator(self, request):
-        auth_user = None
-        if request.user.is_authenticated():
-            auth_user = request.user
-            request.user = AnonymousUser()
-        try:
-            result = f(self, request)
-        finally:
-            if auth_user is not None:
-                request.user = auth_user
-        return result
-    return decorator
 
 class CMSNamedMenuAdmin(admin.ModelAdmin):
+    
     change_form_template = 'cms_named_menus/change_form.html'
 
     readoly_fields = ('pages_json',)
 
     def change_view(self, request, object_id, form_url='', extra_context={}):
-
         menu_pages = CMSNamedMenu.objects.get(id=object_id).pages
-
         extra_context = {
             'menu_pages': json.dumps(menu_pages),
             'available_pages': self.serialize_navigation(request),
             'debug': settings.DEBUG,
         }
+        return super(CMSNamedMenuAdmin, self).change_view(request, object_id, form_url, extra_context)
 
-        return super(CMSNamedMenuAdmin, self).change_view(request,
-                                                          object_id,
-                                                          form_url,
-                                                          extra_context)
-
-    @anonymous_request
     def serialize_navigation(self, request):
-
-        nodes = menu_pool.get_nodes(request)
-
+        nodes = get_nodes(request)
         cleaned = []
-
         for node in nodes:
             node.children = None
             node.parent = None
             cleaned.append(node.__dict__)
-
         return json.dumps(cleaned, cls=LazyEncoder)
